requirements.txt
Django==2.1.1
django-cors-middleware==1.3.1
django-extensions==2.1.2
djangorestframework==3.8.2
PyJWT==1.4.2
six==1.10.0
django-cors-headers==2.4.0
python-decouple==3.1
psycopg2-binary==2.7.5
gunicorn==19.9.0
django-heroku==0.3.1
dj-database-url==0.5.0
django-rest-swagger==2.2.0
coreapi==2.3.3
sendgrid==5.6.0
django-taggit==0.23.0
django-taggit-serializer==0.1.7